Fix GetProduct standard method

+++ backend/backend/Models/Product.cs
@@ -3,7 +3,7 @@ namespace backend.Models;
 public class Product
 {
     public Product(
-        int productId,
+        long productId,
         string productName,
         decimal productPrice,
         ProductCategory productCategory
@@ -15,7 +15,7 @@ public class Product
         ProductCategory = productCategory;
     }
 
-    public int ProductId { get; private set; }
+    public long ProductId { get; private set; }
     public string ProductName { get; private set; }
     public decimal ProductPrice { get; private set; }
     public ProductCategory ProductCategory { get; private set; }
